+++ docs/release-notes.rst
@@ -12,7 +12,9 @@ Release Notes
 * New generator: GNU Autoconf
 * New generator: NuGet
 * New generator: PEAR/PECL
-* New generator: Setuptools
+* New generator: Python Setuptools
+* New generator: Python Wheel
+* New CLI option to skip slow generators
 * Add ``kde-invent`` remote-id
 * Trim ".git" suffix when extracting remote-id